requirements.txt
fastapi
uvicorn[standard]
lancedb
ollama
pyarrow
pydantic
python-dotenv
pandas 